+++ src/App.js
@@ -3,8 +3,7 @@ import { theme } from "./common/Theme/theme";
 import { Header } from "./layouts/Header/index";
 import { useEffect, useState } from "react";
 import { useDispatch } from "react-redux";
-import { fetchMovies } from "./features/movies/movie/moviesSlice";
-import { Content } from "./layouts/Content/index";
+import { MoviesDetails } from "./layouts/MoviesDetails/index";
 
 function App() {
   const dispatch = useDispatch();
@@ -20,14 +19,10 @@ function App() {
     return () => window.removeEventListener("resize", handleResize);
   }, []);
 
-  useEffect(() => {
-    dispatch(fetchMovies());
-  }, []);
-
   return (
     <ThemeProvider theme={theme}>
       <Header />
-      <Content />
+      <MoviesDetails></MoviesDetails>
     </ThemeProvider>
   );
 }
