+++ src/Tile/Description/styled.js
