+++ src/Tile/Details/styled.js
