+++ src/Tile/Information/styled.js
