+++ src/Tile/Raiting/index.js
@@ -1,14 +1,14 @@
 import { MaxRate, RaitingContainer, Rate, StarIcon, Votes } from "./styled";
 
-export const Raiting = ({ movieDetails, vote_average, vote_count }) => {
+export const Raiting = ({ movieDetails }) => {
   return (
     <RaitingContainer movieDetails={movieDetails}>
       <StarIcon alt="" />
       <Rate movieDetails={movieDetails}>
-        {vote_average}
+        7,8
         {movieDetails ? <MaxRate movieDetails={movieDetails}>/ 10</MaxRate> : null}
       </Rate>
-      <Votes movieDetails={movieDetails}>{vote_count} votes</Votes>
+      <Votes movieDetails={movieDetails}>35 votes</Votes>
     </RaitingContainer>
   );
 };
