+++ src/Tile/Raiting/styled.js
